requirements.txt
chainlit>=2.0.0
google-genai>=0.6.0
python-dotenv
